+++ sdk/hardwaresecuritymodules/arm-hardwaresecuritymodules/src/azureDedicatedHSMResourceProvider.ts
@@ -45,15 +45,12 @@ export class AzureDedicatedHSMResourceProvider extends coreClient.ServiceClient 
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-hardwaresecuritymodules/1.0.0`;
+    const packageDetails = `azsdk-js-arm-hardwaresecuritymodules/1.0.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
-    if (!options.credentialScopes) {
-      options.credentialScopes = ["https://management.azure.com/.default"];
-    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,