+++ sdk/hdinsight/arm-hdinsight/src/hDInsightManagementClient.ts
@@ -81,9 +81,6 @@ export class HDInsightManagementClient extends coreClient.ServiceClient {
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
-    if (!options.credentialScopes) {
-      options.credentialScopes = ["https://management.azure.com/.default"];
-    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,