+++ sdk/hybridkubernetes/arm-hybridkubernetes/src/connectedKubernetesClient.ts
@@ -45,15 +45,12 @@ export class ConnectedKubernetesClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-hybridkubernetes/2.0.1`;
+    const packageDetails = `azsdk-js-arm-hybridkubernetes/2.0.2`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
-    if (!options.credentialScopes) {
-      options.credentialScopes = ["https://management.azure.com/.default"];
-    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,