+++ sdk/keyvault/arm-keyvault/src/keyVaultManagementClient.ts
@@ -66,15 +66,12 @@ export class KeyVaultManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-keyvault/2.1.0-beta.1`;
+    const packageDetails = `azsdk-js-arm-keyvault/2.1.0-beta.2`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
-    if (!options.credentialScopes) {
-      options.credentialScopes = ["https://management.azure.com/.default"];
-    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,